ragchatbot/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
cohere==5.11.0
qdrant-client==1.7.0
python-dotenv==1.0.0


